fix: save subsections when starting new section (Bug #6)
Critical parser bug: when encountering a new ## section, the parser
was not saving the current ### subsection to the previous section
before pushing it to the sections array.

This caused ALL subsections to be lost!

Example:
## Section A
### Subsection A1
## Section B  ← Bug: A1 was lost here!

Fixed by saving currentSubsection before pushing currentSection.

This is why ### 关键术语 was being deleted - it was never parsed
into the section in the first place.